refactor(git): move git autentification logic to git auth
